fraudguard-backend/requirements.txt
# Backend Requirements for FraudGuard
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
redis
celery[redis]
pydantic-settings
pytest
httpx
numpy
pandas
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
transformers>=4.35.0
torch>=2.0.0
evaluate>=0.4.0
accelerate>=0.20.0
